+++ src/dexterity/localroles/browser/settings.py
@@ -1,5 +1,6 @@
 # encoding: utf-8
 
+from ast import literal_eval
 from collective.z3cform.datagridfield.row import DictRow
 from copy import deepcopy
 from dexterity.localroles import _
@@ -16,6 +17,7 @@ from dexterity.localroles.browser.interfaces import IRole
 from dexterity.localroles.browser.interfaces import IWorkflowState
 from dexterity.localroles.browser.overrides import CustomTypeFormLayout
 from dexterity.localroles.interfaces import ILocalRolesRelatedSearchUtility
+from imio.helpers.utils import is_valid_json
 from persistent.mapping import PersistentMapping
 from plone import api
 from plone.app.dexterity.interfaces import ITypeSchemaContext
@@ -36,6 +38,8 @@ from zope.component import getUtilitiesFor
 from zope.component import getUtility
 from zope.interface import implementer
 from zope.interface import Interface
+
+import json
 
 
 try:
@@ -118,7 +122,7 @@ class RelatedFormatValidator(validator.SimpleFieldValidator):
         if not value or not value.strip():
             return
         try:
-            var = eval(value)
+            var = literal_eval(value)
         except Exception:
             raise RelatedFormatError
         if not isinstance(var, dict):
@@ -147,7 +151,9 @@ class ILocalRole(Interface):
         required=True,
     )
 
-    related = schema.Text(title=_(u"related role configuration"), required=False)
+    related = schema.Text(title=_(u'related role configuration'),
+                          required=False,
+                          constraint=is_valid_json)
 
 
 validator.WidgetValidatorDiscriminators(RelatedFormatValidator, field=ILocalRole["related"])
@@ -175,7 +181,7 @@ class LocalRoleConfigurationAdapter(object):
             setattr(self.context.fti, "localroles", PersistentMapping())
         old_value = self.context.fti.localroles.get(name, {})
         # for plone 6 with state as tuple
-        for dic in value:
+        for dic in value or []:
             if isinstance(dic["state"], tuple):
                 dic["state"] = dic["state"][0]
         new_dict = self.convert_to_dict(value)
@@ -187,9 +193,9 @@ class LocalRoleConfigurationAdapter(object):
     @staticmethod
     def convert_to_dict(value):
         value_dict = {}
-        for row in value:
+        for row in value or []:
             state, roles, principal = row["state"], row["roles"], row["value"]
-            related = row["related"] is not None and row["related"].strip() and str(eval(row["related"])) or ""
+            related = row["related"] is not None and row["related"].strip() and json.dumps(literal_eval(row["related"])) or ""
             if state not in value_dict:
                 value_dict[state] = {}
             value_dict[state][principal] = {"roles": roles, "rel": related}